validate att rule inputs

+++ src/modules/attendance-rule/attendance-rule.service.ts
@@ -8,25 +8,17 @@ export class AttendanceRuleService {
   constructor(private prisma: PrismaService) {}
 
   async create(rule: AttendanceRuleDto) {
-    const earlyMinute = rule.earlyMinute;
-    const lateMinute = rule.lateMinute;
-    const offDutyTime = rule.offDutyTime;
-    const onDutyTime = rule.onDutyTime;
-
-    if (!Number(earlyMinute) || !Number(lateMinute)) {
-      return 'earlyMinute and lateMinute must be a number';
+    const valid = await this.validateRule(rule);
+    if (valid.status === true) {
+      const exist = await this.prisma.attendanceRule.findFirst();
+      const avail = await this.prisma.attendanceRule.findMany();
+      if (exist && avail.length > 0) {
+        return this.update(exist.id, rule);
+      } else {
+        return await this.prisma.attendanceRule.create({ data: { ...rule } });
+      }
     }
-
-    if (!this.hasTimeFormat(offDutyTime) || !this.hasTimeFormat(onDutyTime)) {
-      return 'offDutyTime or onDutyTime must be a valid time';
-    }
-    const exist = await this.prisma.attendanceRule.findFirst();
-    const avail = await this.prisma.attendanceRule.findMany();
-    if (exist && avail.length > 0) {
-      return this.update(exist.id, rule);
-    } else {
-      return await this.prisma.attendanceRule.create({ data: { ...rule } });
-    }
+    return valid.message;
   }
 
   hasTimeFormat(str: string): boolean {
@@ -43,6 +35,10 @@ export class AttendanceRuleService {
   }
 
   async update(id: number, rule: UpdateAttendanceRuleDto) {
+    const valid = await this.validateRule(rule);
+    if (valid.status === false) {
+      return valid.message;
+    }
     return await this.prisma.attendanceRule.update({
       where: { id },
       data: { ...rule },
@@ -50,10 +46,52 @@ export class AttendanceRuleService {
   }
 
   async deleteOne(id: number) {
+    const exists = await this.prisma.attendanceRule.findUnique({
+      where: { id },
+    });
+    if (!exists) {
+      return 'Attendance Rule does not exist';
+    }
     return await this.prisma.attendanceRule.delete({ where: { id } });
   }
 
   async deleteAll() {
     return await this.prisma.attendanceRule.deleteMany();
+  }
+
+  async validateRule(dto: UpdateAttendanceRuleDto) {
+    const earlyMinute = dto.earlyMinute;
+    const lateMinute = dto.lateMinute;
+    const offDutyTime = dto.offDutyTime;
+    const onDutyTime = dto.onDutyTime;
+
+    if (earlyMinute === '' || lateMinute === '') {
+      return {
+        message: 'earlyMinute and lateMinute must not be empty',
+        status: false,
+      };
+    }
+
+    if (typeof earlyMinute !== 'string' || typeof lateMinute !== 'string') {
+      return {
+        message: 'earlyMinute and lateMinute must be a number in a string',
+        status: false,
+      };
+    }
+
+    if (!Number(earlyMinute) || !Number(lateMinute)) {
+      return {
+        message: 'earlyMinute and lateMinute must be a number',
+        status: false,
+      };
+    }
+
+    if (!this.hasTimeFormat(offDutyTime) || !this.hasTimeFormat(onDutyTime)) {
+      return {
+        message: 'offDutyTime or onDutyTime must be a valid time',
+        status: false,
+      };
+    }
+    return { message: 'valid', status: true };
   }
 }
